Integrate Mask3D/Attn-IPA training from doloresgarcia/maskIPA
Port the maskIPA snapshot (github.com/doloresgarcia/maskIPA, single squashed
commit, no shared history) onto the delphi-dataset branch:

- Add src/models/Mask3D/ (Attn-IPA model, decoder, matcher, losses, kmeans
  cross-attention), inference_oc_mask3d, shower matching utils, windowed
  GATr variants and the example_mode_attn_ipa network config.
- Take upstream-only updates: train_utils (fetch-by-files/train-batches),
  inference metrics, energy_correction_NN, load_pretrained_models.
- Three-way merge (base = fork point) for files changed on both sides:
  object_cond gains allpair-repulsion knobs alongside s_B/s_B_track,
  Gatr_pf_e_noise keeps DELPHI beta-noise weights and adds the
  autocast-off loss wrapper and output_dim passthrough, functions_particles
  keeps the cm->mm rescale on top of the fast awkward->tensor conversion.
- Fix broken imports in the snapshot: drop nonexistent CML_loss and
  dpc_track_seeded modules (only used in dead code).
- Keep our example_mode_gatr_noise wrapper pointing at
  src/models/GATr/Gatr_pf_e_noise.py (upstream rewired it to a cleaned
  top-level copy that lacks the DELPHI patches) and keep our
  loss_regression.py (superset: PID_4_class and legacy encodings).

Both entrypoints verified in the gatr:v9 container: all modules import,
the README mask3d command parses, Attn-IPA builds (15.6M params) and the
DELPHI OC model builds unchanged.

+++ src/logger/plotting_tools.py
@@ -48,8 +48,12 @@ def PlotCoordinates(
             features = torch.sigmoid(graph_i.ndata["beta"])
 
         tidx = graph_i.ndata["particle_number"]
-        # .float(): numpy cannot represent bfloat16 (model outputs under bf16-mixed)
+        coords = coords.float()
+        features = features.float()
         data = {
+            # .float() upcasts: numpy has no bfloat16, so under --use-amp
+            # (bf16-mixed) coords/features are BFloat16 and .numpy() raises
+            # "Got unsupported ScalarType BFloat16". No-op for fp32.
             "X": coords[:, 0].view(-1, 1).detach().float().cpu().numpy(),
             "Y": coords[:, 1].view(-1, 1).detach().float().cpu().numpy(),
             "Z": coords[:, 2].view(-1, 1).detach().float().cpu().numpy(),
